Wrap both TalonSRXs and SparkMAXes into one interface
I went ahead and introduced a "SmartMotorController" interface, then
implemented it in two new classes: "MetroSparkMAX", and "MetroTalonSRX",
I'm not sure if I added everything we use but it's pretty easy to append
to. Hopefully this will be pretty easy to hook into if we need to switch
everything to Falcons or whatever in the future.