Change MotorSubsystem Map argument to List (#16)
* Moved MotorSubsystem Map to List

Moved MotorSubsystem Map to a List. This way, there are less strings and
hopefully everything makes more sense. MotorCommand now optionally takes
an index if only one motor needs to be moved, otherwise it will move all
of them in the list.

* Change Intake Commands

Build failed because the intake commands still used the Map calls to
setSpeed. Those commands arent used anywhere, but I updated them anyway
for convenience.

* Make MotorSubsystem and MotorCommand More Readable

Changed the shortened null-check if statement in MotorCommand to a
multiline one for readability, and changed the for loop in the init
block of MotorSubsystem to a .forEach for consistency within the class.