Port Pivot to Library (#18)
* Add Intake and Pivot Motors to Consts

Added the motor for intake and the motor for pivot to Consts.kt so as to
put all of the motor declarations in the same place. Also changed
Intake.kt to use the motor from consts instead of declaring its own.

* Port Pivot to Library

Changed Pivot to extend MotorSubsystem, and simplified its contents.
Removed usage of Pivot-specific commands and replaced them with
MotorCommand and PIDCommands.

* Remove Usages of RunPivot

Removed a useage of RunPivot that I missed originally, and replaced it
with a MotorCommand.

* Replace Broken PIDCommands with MotorCommands

Replaced the PIDCommands used for controlling Pivot (which did not work)
with more generic MotorCommands. The values 0.5 and -0.5 are
placeholders, and can be changed if they're too fast/slow.