Merge Library (#19)
* Added Library Commands/Subsystems

Added MotorCommand, PIDCommand, and PneumaticShiftCommand to the new
library directory. These are not integrated into the various subsystems yet. New commands/subsystems will be added to the library as necessary.

* Moved Util Stuff to Library

Moved PIDCommand, ToggleLightCommand, and ToggleRelayCommand (previously SwitchRelay) from robot/util to library/commands. Name changes were made to keep consistancy and avoid confusion.

* Moved Climber to MotorSubsystem

Moved the climber object to MotorSubsystem class instead of Climber
class. Climber still exists in the code for reference, but isn't used
anywhere. MotorSubsystem now is constructed with a map for multiple
motors and their names.

* Change toggles to use when statements

More idiomatic code, and reads a bit nicer. Also minimizes calls to
"get" functions which might or might not be expensive with regards to
performance.

* Move physics stuff to library

Move physics resources to library since we'll probably keep using them.

* Fix compile

* Ported Intake to Library

Changed robot/intake/Intake.kt to extend MotorSubsystem, reducing
redundant code (Issue #5). Moved all intake commands in RobotContainer.kt to MotorCommands. Changed all intake commands to use the new Intake subsystem, in case they still need to be used. Tweaked MotorSubsystem for better use when it's being inherited from.

* Wrap both TalonSRXs and SparkMAXes into one interface

I went ahead and introduced a "SmartMotorController" interface, then
implemented it in two new classes: "MetroSparkMAX", and "MetroTalonSRX",
I'm not sure if I added everything we use but it's pretty easy to append
to. Hopefully this will be pretty easy to hook into if we need to switch
everything to Falcons or whatever in the future.

* Change 'this' to super where applicable for clarity

* Minor clean-up and testing

Testing the library today, so far just tested intake and everything works, I went ahead and did some minor cleanup.

* Change Consts to motors instead of ports

Using it this way introduces some dangers with public motor controller classes but also works better for our dependency injection-esque way of doing things. I'm down for changes later.

* Move Storage to Library (#15)

* Changed storage to an instance of MotorSubsystem

* Changed RunStorageConstant to MotorCommand

* Fixed Storage Implementation of MotorSubsystem

Changed storage's MotorSubsystem constructor to use motors in the
constants instead of instantiation them inside the constructor. Also
fixed missing commas in some MotorCommand calls, as well as moving
FinalAutoGroup to use MotorSubsystem for storage instead of Storage()
class. FinalAutoGroup (and other auto commands) will need generalized
further.

Co-authored-by: Andrew <kandrew2022@gmail.com>

* Change MotorSubsystem Map argument to List (#16)

* Moved MotorSubsystem Map to List

Moved MotorSubsystem Map to a List. This way, there are less strings and
hopefully everything makes more sense. MotorCommand now optionally takes
an index if only one motor needs to be moved, otherwise it will move all
of them in the list.

* Change Intake Commands

Build failed because the intake commands still used the Map calls to
setSpeed. Those commands arent used anywhere, but I updated them anyway
for convenience.

* Make MotorSubsystem and MotorCommand More Readable

Changed the shortened null-check if statement in MotorCommand to a
multiline one for readability, and changed the for loop in the init
block of MotorSubsystem to a .forEach for consistency within the class.

* Port Pivot to Library (#18)

* Add Intake and Pivot Motors to Consts

Added the motor for intake and the motor for pivot to Consts.kt so as to
put all of the motor declarations in the same place. Also changed
Intake.kt to use the motor from consts instead of declaring its own.

* Port Pivot to Library

Changed Pivot to extend MotorSubsystem, and simplified its contents.
Removed usage of Pivot-specific commands and replaced them with
MotorCommand and PIDCommands.

* Remove Usages of RunPivot

Removed a useage of RunPivot that I missed originally, and replaced it
with a MotorCommand.

* Replace Broken PIDCommands with MotorCommands

Replaced the PIDCommands used for controlling Pivot (which did not work)
with more generic MotorCommands. The values 0.5 and -0.5 are
placeholders, and can be changed if they're too fast/slow.

Co-authored-by: Mani Kamali <kamalisarvestani.1@themetroschool.org>
Co-authored-by: Mani Kamali <kamalimani1623@gmail.com>
Co-authored-by: maniyar1 <32717947+maniyar1@users.noreply.github.com>
Co-authored-by: Team Laptop <team3324metrobots@gmail.com>
Co-authored-by: treeden <37743981+treeden@users.noreply.github.com>